[Assignment Tracking][NFC] Use BitVectors as masks for SmallVectors rather than using DenseMaps
Rather than tracking 3 maps of `{VariableID: SomeInfo}` per block, use a `BitVector` indexed by `VariableID` to mask 3 vectors of `SomeInfo`.

`BlockInfo`s now need to be initialised with a call to `init` which sets the `BitVector` width to the number of partially promoted variables in the function and fills the vectors with Top values.

Prior to this patch, in `joinBlockInfo`, it was necessary to insert Top values into the `Join` result for variables in `A XOR B` after joining the variables in `A AND B`. Now, because the vectors are pre-filled with Top values we need only join the variables `A AND B` and set the `BitVector` of tracked variables to `A OR B`.

Differential Revision: https://reviews.llvm.org/D145558