[DebugInfo][RemoveDIs] Reverse order of DPValues from findDbgUsers
The order of dbg.value intrinsics appearing in the output can affect the
order of tables in DWARF sections. This means that DPValues, our dbg.value
replacement, needs to obey the same ordering rules. For dbg.values returned
by findDbgUsers it's reverse order of creation (due to how they're put on
use-lists). Produce that order from findDbgUsers for DPValues.