[KeyInstr][CC1] Add cc1 option -g[no-]key-instructions
This controls the work-in-progress feature Key Instructions. Once it's no longer
a WIP we can look at making this a Clang option.

The LLVM patch stack starts at #133477.

The feature is only functional in LLVM if LLVM is built with CMake flag
LLVM_EXPERIMENTAL_KEY_INSTRUCTIONs. Eventually that flag will be removed.

The Key Instructions feature is discussed here:
https://discourse.llvm.org/t/rfc-improving-is-stmt-placement-for-better-interactive-debugging/82668

The Clang-side work is demoed here:
https://github.com/llvm/llvm-project/pull/130943