[NFC][AssignmentTracking] Insert locs rather than append
In the second phase of the analysis extra locations are generated to ensure that
memory locations are not invalided by non-memory locations for part of a
variable.

These were appended to the running "insert-before-position" vectors. Change this
to insert the new locations into the vectors rather than append.

This lets us get rid of some insertion-location complexities in later patches,
as we no longer have to track the position in IR